increase robustness of full-elan script in case of phase1 failure

+++ bin/control/go-full-elan.sh
@@ -9,30 +9,37 @@ curl -X POST -H 'Content-type: application/json' --data "$MSG" $SLACK_MGMT_HOOK
 
 DATESTAMP=`date '+%Y%m%d'`
 
-$NEXTFLOW_BIN run elan.nf -c elan.config --dump $PRE_ELAN_DIR/latest.tsv --publish $ELAN_DIR --schemegit /cephfs/covid/software/sam/artic-ncov2019 --datestamp $DATESTAMP > nf.elan.$DATESTAMP.log 2>&1;
-ret=$?
-mv .nextflow.log elan.nextflow.log
-
-if [ $ret -ne 0 ]; then
-    lines=`tail -n 25 nf.elan.$DATESTAMP.log`
-else
-    lines=`awk -vRS= 'END{print}' nf.elan.$DATESTAMP.log`
-fi
-
-MSG='{"text":"*COG-UK inbound pipeline finished...*
-...with exit status '"$ret"'
-'"\`\`\`${lines}\`\`\`"'
- _Have a nice day!_"}'
-curl -X POST -H 'Content-type: application/json' --data "$MSG" $SLACK_MGMT_HOOK
-
-if [ $ret -ne 0 ]; then
-    MSG='{"text":"<!channel> *COG-UK inbound pipeline failed...*"}'
-    curl -X POST -H 'Content-type: application/json' --data "$MSG" $SLACK_MGMT_HOOK
-    exit $ret
-fi
-
 # OCARINA_FILE only written if elan processed at least one sample
 OCARINA_FILE="$ELAN_DIR/staging/summary/$DATESTAMP/ocarina.files.ls"
+
+if [ ! -f "$OCARINA_FILE" ]; then
+    $NEXTFLOW_BIN run elan.nf -c elan.config --dump $PRE_ELAN_DIR/latest.tsv --publish $ELAN_DIR --schemegit /cephfs/covid/software/sam/artic-ncov2019 --datestamp $DATESTAMP > nf.elan.$DATESTAMP.log 2>&1;
+    ret=$?
+    mv .nextflow.log elan.nextflow.log
+
+    if [ $ret -ne 0 ]; then
+        lines=`tail -n 25 nf.elan.$DATESTAMP.log`
+    else
+        lines=`awk -vRS= 'END{print}' nf.elan.$DATESTAMP.log`
+    fi
+
+    MSG='{"text":"*COG-UK inbound pipeline finished...*
+...with exit status '"$ret"'
+'"\`\`\`${lines}\`\`\`"'
+_Have a nice day!_"}'
+    curl -X POST -H 'Content-type: application/json' --data "$MSG" $SLACK_MGMT_HOOK
+
+    if [ $ret -ne 0 ]; then
+        MSG='{"text":"<!channel> *COG-UK inbound pipeline failed...*"}'
+        curl -X POST -H 'Content-type: application/json' --data "$MSG" $SLACK_MGMT_HOOK
+        exit $ret
+    fi
+else
+    MSG='{"text":"*COG-UK inbound pipeline* Cowardly skipping Elan as the ocarina file already exists for today"}'
+    curl -X POST -H 'Content-type: application/json' --data "$MSG" $SLACK_MGMT_HOOK
+fi
+
+# If the OCARINA_FILE has still not been written at this point, it means Elan ran but today's pipeline is empty - abort early but successfully, dont send a tael
 if [ ! -f "$OCARINA_FILE" ]; then
     curl -X POST -H 'Content-type: application/json' --data '{"text":"\n*COG-UK inbound pipeline empty*\nNo new valid files today, try again tomorrow."}' $SLACK_REAL_HOOK
     exit 0
